Move shared functions out of location.py
Moves :func:`load_input_from_file_or_stdin`,
:func:`drop_columns_from_output`, and :class:`ColumnDoesNotExistError`
to shared module. They will also be used by the de-identify command.
